feat: fix list element mapping and add comprehensive test coverage
Generate list.map { } calls in IR when list element types differ,
fixing a bug where List<SourceItem> → List<TargetItem> passed elements
through without transformation. Also adds missing test cases from the
implicit mappings spec (widening + value class edge cases) and splits
the monolithic KMapperTest into 6 focused test files.
